% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_one.R
\name{create_table_one}
\alias{create_table_one}
\title{Simulation Table 1}
\usage{
create_table_one(K_f = 4, L_f = 4, K = 4, L = 4, J = c(50, 50,
  50), K_true, R = 1, f_s = c(function(x) 3 * ((1/2) * dbeta(x/R, 1, 8)
  + (1/2) * dbeta(x/R, 6, 1)), function(x) 3 * ((1/5) * dbeta(x/R, 3, 2) +
  (2/3) * dbeta(x/R, 3, 1) + (2/15) * dbeta(x/R, 1, 1)), function(x) 3 *
  ((1/2) * dbeta(x/R, 8, 2) + (1/2) * dbeta(x/R, 30, 50))),
  iter_max = 15000, warm_up = 1.3e+42, thin = 1)
}
\arguments{
\item{K_f}{number of cluster components in fhmm}

\item{L_f}{number of within-cluster mixture components in fhmm}

\item{K}{number of cluster sticks for NDP}

\item{L}{number of within-cluster mixture-sticks for NDP}

\item{J}{total number of groups}

\item{K_true}{true number of clusters}

\item{f_s}{list of functions that should be the same length as K_true}
}
\description{
This function creates a table comparing the novel NDP
to a classic heirarchical finite mixture model
}
